agrego envio de porcentaje de cpu a planificador y recepcion, falta limpiar el porcentaje sino hace nada

+++ CPU/CPU.h
@@ -72,7 +72,7 @@ typedef struct {
 	int id;
 	int socket;
 	int pid;
-	int porcentajeUso;
+	double porcentajeUso;
 } t_cpu;
 
 typedef struct __attribute__((packed))
